requirements.txt
streamlit==1.26.0
nltk==3.8.1
scikit-learn==1.2.2
numpy==1.24.4
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
Pillow==9.5.0
